+++ Providers/CrashlyticsProvider.m
@@ -35,14 +35,6 @@
     CLSLog(@"%@", event);
 }
 
-- (void)didShowNewViewController:(UIViewController *)controller {
-    [self event:@"Screen view" withProperties:@{ @"screen": controller.title }];
-}
-
-- (void)logTimingEvent:(NSString *)event withInterval:(NSNumber *)interval {
-    [self event:event withProperties:@{ @"length": interval }];
-}
-
 - (void)remoteLog:(NSString *)parsedString {
     CLSLog(@"%@", parsedString);
 }