+++ Providers/CrittercismProvider.h
@@ -9,5 +9,4 @@
 #import "ARAnalyticalProvider.h"
 
 @interface CrittercismProvider : ARAnalyticalProvider
-
 @end