+++ Providers/FlurryProvider.h
@@ -9,5 +9,4 @@
 #import "ARAnalyticalProvider.h"
 
 @interface FlurryProvider : ARAnalyticalProvider
-
 @end