+++ Providers/KISSMetricsProvider.h
@@ -9,5 +9,4 @@
 #import "ARAnalyticalProvider.h"
 
 @interface KISSMetricsProvider : ARAnalyticalProvider
-
 @end