+++ Providers/LocalyticsProvider.h
@@ -9,5 +9,4 @@
 #import "ARAnalyticalProvider.h"
 
 @interface LocalyticsProvider : ARAnalyticalProvider
-
 @end