+++ Providers/MixpanelProvider.m
@@ -38,13 +38,5 @@
     [[Mixpanel sharedInstance] track:event properties:properties];
 }
 
-- (void)didShowNewViewController:(UIViewController *)controller {
-    [self event:@"Screen view" withProperties:@{ @"screen": controller.title }];
-}
-
-- (void)logTimingEvent:(NSString *)event withInterval:(NSNumber *)interval {
-    [self event:event withProperties:@{ @"length": interval }];
-}
-
 #endif
 @end