+++ Providers/TestFlightProvider.h
@@ -9,7 +9,6 @@
 #import <Foundation/Foundation.h>
 #import "ARAnalyticalProvider.h"
 
-
 @interface TestFlightProvider : ARAnalyticalProvider 
 @end