+++ Providers/TestFlightProvider.m
@@ -48,14 +48,6 @@
     [TestFlight passCheckpoint:event];
 }
 
-- (void)didShowNewViewController:(UIViewController *)controller {
-    [self event:@"Screen view" withProperties:@{ @"screen": controller.title }];
-}
-
-- (void)logTimingEvent:(NSString *)event withInterval:(NSNumber *)interval {
-    [self event:event withProperties:@{ @"length": interval }];
-}
-
 - (void)remoteLog:(NSString *)parsedString {
     TFLog(parsedString);
 }