+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 #ARAnalytics
 
+## Version 1.1
+
+* Support for Countly
+* Rewrite into a more modular codebase
+* Supports more features of Google Analytics 
+
 ## Version 1.0
 
 * Supports TestFlight, Mixpanel, Localytics, Flurry, Google Analytics, KISSMetrics, Crittercism and Crashlytics